refactor(desktop): extract main window controller
