+++ apps/desktop/src/main/main.ts
@@ -1,6 +1,4 @@
-import { app, BrowserWindow, dialog, ipcMain, Menu, nativeImage, nativeTheme, safeStorage, screen, shell } from 'electron';
-import { isExternalUrl } from './external-link-guard.js';
-import { readSavedBounds, writeSavedBounds, type SavedBounds } from './window-state.js';
+import { app, ipcMain, nativeImage, safeStorage, shell } from 'electron';
 import { randomUUID } from 'node:crypto';
 import { copyFile, mkdir, readFile, realpath } from 'node:fs/promises';
 import { isAbsolute, join, relative, resolve, sep } from 'node:path';
@@ -242,12 +240,11 @@ import {
   persistSynthesisCacheBlocksToArtifacts,
 } from './synthesis-cache-artifacts.js';
 import { buildBrowserTools } from './browser/browser-tools.js';
-import { BrowserViewManager } from './browser/view-manager.js';
-import { BrowserViewController } from './browser/controller.js';
 import { createBrowserViewHost } from './browser/automation-host.js';
 import { provideBrowserViewHost } from './browser/browser-host.js';
 import { releaseBrowserSession, revokeHiddenBrowserActions } from './browser/session.js';
 import type { BrowserViewRect } from './browser/logic.js';
+import { createMainWindowController } from './main-window.js';
 
 const buildInfo = resolveBuildInfo(app.isPackaged, app.getAppPath());
 
@@ -576,6 +573,13 @@ const localMemory = new LocalMemoryService({
   updateSettings: (patch) => settingsStore.update(patch),
   getPrivacyContext: getWorkspacePrivacyContext,
 });
+const mainWindowController = createMainWindowController({
+  workspaceRoot,
+  visualSmokeFixture,
+  settingsStore,
+  ensureBundledOfficeSkills,
+});
+const safeSendToRenderer = mainWindowController.send;
 const openGateway = new OpenGatewayService({
   getSettings: () => settingsStore.get(),
   listSessions: () => runtime.listSessions(),
@@ -740,9 +744,7 @@ async function saveMarkdownViaDialog(
     defaultPath: safeName,
     filters: [{ name: 'Markdown', extensions: ['md'] }],
   };
-  const result = mainWindow
-    ? await dialog.showSaveDialog(mainWindow, saveDialogOptions)
-    : await dialog.showSaveDialog(saveDialogOptions);
+  const result = await mainWindowController.showSaveDialog(saveDialogOptions);
   if (result.canceled || !result.filePath) return { ok: false, reason: 'canceled' };
   try {
     const { writeFile } = await import('node:fs/promises');
@@ -1293,38 +1295,11 @@ const onboardingService = createOnboardingService(
   }),
 );
 
-let mainWindow: BrowserWindow | null = null;
-
-/**
- * Guarded `webContents.send` for `mainWindow`. The `mainWindow?.` optional
- * chain only covers a null reference — it does NOT catch the case where the
- * BrowserWindow has been destroyed (window closed, renderer crashed,
- * teardown raced) while the variable still points at the freed object.
- * Calling `.webContents.send` in that state throws `TypeError: Object has
- * been destroyed`, surfacing as a main-process JS-error dialog.
- *
- * Use this helper anywhere a timer / IPC / menu accelerator might race
- * window teardown. No-op when the window is gone — callers that need
- * delivery confirmation should observe their own state.
- */
-function safeSendToRenderer(channel: string, ...args: unknown[]): void {
-  if (!mainWindow || mainWindow.isDestroyed()) return;
-  const wc = mainWindow.webContents;
-  if (wc.isDestroyed()) return;
-  wc.send(channel, ...args);
-}
-
-const MAIN_WINDOW_TRAFFIC_LIGHT_POSITION = { x: 14, y: 14 } as const;
-const HIDDEN_TRAFFIC_LIGHT_POSITION = { x: -100, y: -100 } as const;
 const planReminderTimers = new Map<string, NodeJS.Timeout>();
 const PLAN_REMINDER_DEFAULT_SNOOZE_MS = 10 * 60 * 1000;
 let dailyReviewSchedulerTimer: NodeJS.Timeout | null = null;
 let dailyReviewSchedulerLastMinuteKey: string | null = null;
 
-// Embedded browser: one WebContentsView per conversation, lazily created on
-// first use. The factory reads the live mainWindow at create time, so views
-// created after a window re-open attach to the current window.
-let browserViews: BrowserViewManager<BrowserViewController> | undefined;
 // The session the renderer currently shows; browser:* renderer channels are
 // validated against it so a stale/miswired panel can't steer another
 // conversation's view (the agent path uses the runtime's trusted sessionId).
@@ -1606,351 +1581,6 @@ async function runDailyReviewIfMissing(mode: DailyReviewMode): Promise<void> {
   await runDailyReview({ mode, trigger: 'cron' });
 }
 
-function getBrowserViews(): BrowserViewManager<BrowserViewController> {
-  if (!browserViews) {
-    browserViews = new BrowserViewManager<BrowserViewController>({
-      create: (sessionId) => {
-        if (!mainWindow) throw new Error('Embedded browser used before the window is ready.');
-        return new BrowserViewController(mainWindow, sessionId, (sid, state) => {
-          safeSendToRenderer('browser:state', { sessionId: sid, state });
-        });
-      },
-      onLiveChange: (sessionIds) => safeSendToRenderer('browser:live', { sessionIds }),
-    });
-  }
-  return browserViews;
-}
-
-/**
- * Guard against saved x/y referencing a display that no longer exists
- * (laptop docked → undocked, external monitor unplugged). Walks the
- * current display workAreas; if no display contains a meaningful
- * overlap with the saved bounds, strip x/y so Electron centers the
- * window on the primary display.
- *
- * "Meaningful overlap" = at least a 100×100 corner of the saved
- * rectangle lies inside some display's workArea. Tighter than "any
- * pixel intersects" so a 1px sliver still flagged-as-off-screen
- * doesn't leave a tiny visible nub the user has to grab.
- */
-function clampBoundsToVisibleDisplay(bounds: SavedBounds): SavedBounds {
-  if (bounds.x === undefined || bounds.y === undefined) return bounds;
-  const displays = screen.getAllDisplays();
-  if (displays.length === 0) return { width: bounds.width, height: bounds.height };
-  const visible = displays.some((display) => {
-    const wa = display.workArea;
-    const overlapX = Math.max(0, Math.min(bounds.x! + bounds.width, wa.x + wa.width) - Math.max(bounds.x!, wa.x));
-    const overlapY = Math.max(0, Math.min(bounds.y! + bounds.height, wa.y + wa.height) - Math.max(bounds.y!, wa.y));
-    return overlapX >= 100 && overlapY >= 100;
-  });
-  if (visible) return bounds;
-  // Off-screen: keep the size but drop the position so Electron centers.
-  return { width: bounds.width, height: bounds.height, isMaximized: bounds.isMaximized };
-}
-
-function visualSmokeWindowBounds(defaults: SavedBounds): SavedBounds {
-  if (!visualSmokeFixture) return defaults;
-  const width = Number(process.env.MAKA_VISUAL_SMOKE_WIDTH);
-  const height = Number(process.env.MAKA_VISUAL_SMOKE_HEIGHT);
-  if (
-    Number.isFinite(width) &&
-    Number.isFinite(height) &&
-    width >= 480 &&
-    height >= 320
-  ) {
-    return { width: Math.floor(width), height: Math.floor(height) };
-  }
-  return defaults;
-}
-
-async function createWindow(): Promise<void> {
-  await mkdir(workspaceRoot, { recursive: true });
-  await ensureBundledOfficeSkills(workspaceRoot);
-  installApplicationMenu();
-  // Restore previously-saved bounds when available; first launch and
-  // legacy installs both fall back to the default 1240x820 frame. After
-  // load, validate the saved x/y against the current display layout — if
-  // the previous external monitor is gone, drop x/y so Electron centers
-  // the window on the primary display instead of opening it off-screen.
-  const defaults = visualSmokeWindowBounds({ width: 1240, height: 820 });
-  const savedBounds = visualSmokeFixture
-    ? defaults
-    : await readSavedBounds(workspaceRoot, defaults);
-  const bounds = clampBoundsToVisibleDisplay(savedBounds);
-
-  // @kenji PR103 follow-up: complete the FOUC fix at the window-chrome layer.
-  // The renderer applies `.dark` synchronously before React mounts (PR103),
-  // but the BrowserWindow's `backgroundColor` shows during the first frame
-  // before the renderer paints. Pick the right initial bg by reading the
-  // persisted theme + system preference.
-  // PR-IR-01b: visual smoke theme override wins over the persisted user
-  // pref. This guarantees the BrowserWindow backgroundColor matches the
-  // theme variant we're about to screenshot, so the very first frame
-  // doesn't capture a light-on-dark or dark-on-light flash.
-  const persistedTheme = (await settingsStore.get()).appearance?.theme ?? 'auto';
-  const themePref = visualSmokeFixture?.theme ?? persistedTheme;
-  const isDark =
-    themePref === 'dark' ||
-    (themePref === 'auto' && nativeTheme.shouldUseDarkColors);
-  const initialBg = isDark ? '#1c1d21' : '#f3f3f5';
-
-  mainWindow = new BrowserWindow({
-    width: bounds.width,
-    height: bounds.height,
-    ...(bounds.x !== undefined && bounds.y !== undefined ? { x: bounds.x, y: bounds.y } : {}),
-    title: 'Maka',
-    // PR-GRAY-CARD-LIFT-0 (WAWQAQ msg `0eb99429` 2026-06-20): the
-    // app icon ships as a 1024px PNG under apps/desktop/assets/icon.png.
-    // BrowserWindow accepts a PNG path directly on macOS for the dock
-    // / window title bar; .icns / .ico packaging will come with the
-    // installer build pass. The asset path resolves from the built
-    // dist/main/main.js (two levels up to apps/desktop, then assets).
-    icon: join(import.meta.dirname, '..', '..', 'assets', 'icon.png'),
-    titleBarStyle: 'hiddenInset',
-    trafficLightPosition: MAIN_WINDOW_TRAFFIC_LIGHT_POSITION,
-    // PR-SIDEBAR-IA-0 Phase 3 P0 fixup v5 (WAWQAQ msg `5b85fdb1`,
-    // xuan `eea556cd`): explicit `resizable: true` so a future
-    // patch can't silently disable window edge resize. Default is
-    // already `true`, but pinning it here removes the ambiguity
-    // and makes the intent obvious to reviewers; CSS-level fixes
-    // (see `app-region-hygiene-contract.test.ts`) cover the
-    // renderer side of the same gate.
-    resizable: true,
-    backgroundColor: initialBg,
-    // PR-VISUAL-SMOKE-HEADLESS: under visual-smoke capture, never show the
-    // window or let the app take foreground — captures run while the
-    // developer keeps working in another app. `webContents.capturePage()`
-    // still returns a painted frame on a hidden window because
-    // `paintWhenInitiallyHidden` defaults to true. Real runs keep the
-    // default `show: true`.
-    ...(process.env.MAKA_VISUAL_SMOKE_FIXTURE ? { show: false } : {}),
-    // Glass material — reference-atlas §1 + §12.1 documents the upstream
-    // reference layout's `light-glass` / `dark-glass` themes that paint
-    // the sidebar against native macOS vibrancy material. Enabling
-    // `vibrancy: 'sidebar'` here lets the CSS-side sidebar render
-    // transparent and inherit the system's blurred window material
-    // (Big Sur+). Renderer CSS gates the transparency on
-    // `[data-vibrancy="active"]` so non-macOS builds (where vibrancy is
-    // a no-op) keep their opaque chrome.
-    // Skip vibrancy under MAKA_VISUAL_SMOKE_FIXTURE — capture environments
-    // can't paint native window material reliably, and the auto-capture
-    // renderer would stall waiting for compositor frames that never settle.
-    ...(process.platform === 'darwin' && !process.env.MAKA_VISUAL_SMOKE_FIXTURE
-      ? { vibrancy: 'sidebar' as const }
-      : {}),
-    webPreferences: {
-      preload: join(import.meta.dirname, '..', 'preload', 'preload.cjs'),
-      // Defense-in-depth flags (@kenji PR96 review). The external-link guard
-      // is the perimeter; these settings keep a hostile page from reaching
-      // Node primitives even if it somehow loaded inside the BrowserWindow:
-      contextIsolation: true,    // window.maka via contextBridge only
-      nodeIntegration: false,    // no `require` in renderer
-      sandbox: true,             // preload runs in the renderer sandbox
-      webSecurity: true,         // enforce CSP / same-origin policy
-      allowRunningInsecureContent: false,
-    },
-  });
-
-  // Two-layer external-link hygiene: assistant markdown often emits `<a href>`
-  // links to docs / GitHub / provider sign-up pages. Without these guards
-  // clicking such a link would either replace the renderer view with the
-  // remote page (breaking the app) or open a new BrowserWindow with full
-  // Node integration.
-  //
-  // 1. `setWindowOpenHandler` intercepts `target="_blank"` and JS `window.open`,
-  //    hands the URL to the OS, denies the in-app open.
-  // 2. `will-navigate` blocks plain `<a>` clicks that would replace the
-  //    renderer location with a non-file:// URL, opening externally instead.
-  //
-  // Both are gated on the URL using `http(s):` or `mailto:` — everything else
-  // (file://, electron internal, etc.) is allowed/denied per Electron defaults.
-  mainWindow.webContents.setWindowOpenHandler(({ url }) => {
-    if (isExternalUrl(url)) {
-      void shell.openExternal(url);
-    }
-    return { action: 'deny' };
-  });
-  mainWindow.webContents.on('will-navigate', (event, url) => {
-    // The initial Vite dev-server / packaged file:// load is allowed through
-    // (current URL equals navigation target while the renderer is settling).
-    // Every subsequent navigation is blocked: external URLs (http/https/
-    // mailto) get handed off to the OS, internal/file:// (including dropped
-    // files attempting to navigate to `file:///…`) are dropped entirely so
-    // the renderer never loses its React tree.
-    const current = mainWindow?.webContents.getURL() ?? '';
-    if (current === url) return;
-    event.preventDefault();
-    if (isExternalUrl(url)) {
-      void shell.openExternal(url);
-    }
-  });
-
-  // Block in-window file drops. Without this, dropping a file onto the
-  // BrowserWindow tries to navigate to its `file://` URL; the `will-navigate`
-  // handler above stops the navigation, but the visual flash + dropEffect
-  // ambiguity is still confusing. Suppressing dragover/drop at the document
-  // level keeps the chat surface immutable to accidental drops.
-  mainWindow.webContents.on('did-finish-load', () => {
-    mainWindow?.webContents.executeJavaScript(`
-      (() => {
-        const block = (e) => { e.preventDefault(); e.stopPropagation(); };
-        window.addEventListener('dragover', block, true);
-        window.addEventListener('drop', block, true);
-      })();
-    `).catch(() => { /* renderer may not be ready; ignore */ });
-  });
-
-  // Restore maximized state after construction (BrowserWindow constructor
-  // doesn't accept it directly; calling here keeps the unmaximized bounds
-  // accurate for the next save).
-  if (bounds.isMaximized) {
-    mainWindow.maximize();
-  }
-
-  // Persist bounds across launches. Debounce so a continuous resize drag
-  // doesn't write the file on every frame; flush on close.
-  let saveTimer: NodeJS.Timeout | undefined;
-  const scheduleSave = () => {
-    if (!mainWindow) return;
-    if (saveTimer) clearTimeout(saveTimer);
-    saveTimer = setTimeout(() => {
-      if (!mainWindow) return;
-      const next: SavedBounds = mainWindow.isMaximized()
-        ? { ...mainWindow.getNormalBounds(), isMaximized: true }
-        : { ...mainWindow.getBounds(), isMaximized: false };
-      void writeSavedBounds(workspaceRoot, next);
-    }, 400);
-  };
-  mainWindow.on('resize', scheduleSave);
-  mainWindow.on('move', scheduleSave);
-  mainWindow.on('maximize', scheduleSave);
-  mainWindow.on('unmaximize', scheduleSave);
-  mainWindow.on('close', () => {
-    if (saveTimer) clearTimeout(saveTimer);
-    // The window owns the embedded-browser views (children of its contentView);
-    // tear them down so their WebContents close with it instead of leaking.
-    void browserViews?.disposeAll();
-    if (!mainWindow) return;
-    const final: SavedBounds = mainWindow.isMaximized()
-      ? { ...mainWindow.getNormalBounds(), isMaximized: true }
-      : { ...mainWindow.getBounds(), isMaximized: false };
-    void writeSavedBounds(workspaceRoot, final);
-  });
-
-  if (process.env.VITE_DEV_SERVER_URL) {
-    await mainWindow.loadURL(process.env.VITE_DEV_SERVER_URL);
-  } else {
-    await mainWindow.loadFile(join(import.meta.dirname, '..', 'renderer', 'index.html'));
-  }
-  if (process.env.MAKA_REAL_WINDOW_SMOKE === '1') {
-    emitRealWindowSmokeDiagnostic('after-load');
-    setTimeout(() => emitRealWindowSmokeDiagnostic('settled-1000ms'), 1000);
-  }
-}
-
-function emitRealWindowSmokeDiagnostic(stage: string): void {
-  const target = mainWindow;
-  if (!target) {
-    console.log(`[real-window-smoke] diagnostic ${JSON.stringify({ stage, windowExists: false })}`);
-    return;
-  }
-  const windowState = {
-    stage,
-    windowExists: true,
-    title: target.getTitle(),
-    bounds: target.getBounds(),
-    normalBounds: target.getNormalBounds(),
-    isVisible: target.isVisible(),
-    isFocused: target.isFocused(),
-    isMinimized: target.isMinimized(),
-    isMaximized: target.isMaximized(),
-    isResizable: target.isResizable(),
-    isMovable: target.isMovable(),
-    isModal: target.isModal(),
-    webContentsUrl: target.webContents.getURL(),
-  };
-  target.webContents
-    .executeJavaScript(
-      `(() => ({
-        readyState: document.readyState,
-        title: document.title,
-        appFramePresent: Boolean(document.querySelector('.appFrame')),
-        searchModalPresent: Boolean(document.querySelector('.maka-search-modal')),
-        searchModalBackdropPresent: Boolean(document.querySelector('.maka-dialog-backdrop')),
-        errorBoundaryPresent: Boolean(document.querySelector('.maka-error-surface')),
-        activeElementInSearchModal: Boolean(document.activeElement && document.activeElement.closest && document.activeElement.closest('.maka-search-modal')),
-        activeElement: document.activeElement ? {
-          tagName: document.activeElement.tagName,
-          className: typeof document.activeElement.className === 'string' ? document.activeElement.className : '',
-          ariaLabel: document.activeElement.getAttribute('aria-label'),
-        } : null,
-      }))()`,
-      true,
-    )
-    .then((rendererState) => {
-      console.log(`[real-window-smoke] diagnostic ${JSON.stringify({ ...windowState, renderer: rendererState })}`);
-    })
-    .catch((err: unknown) => {
-      console.log(`[real-window-smoke] diagnostic ${JSON.stringify({ ...windowState, rendererError: errorMessage(err) })}`);
-    });
-}
-
-
-function installApplicationMenu(): void {
-  // App menu labels match the in-app Chinese-leaning UI per the PR69/70/71
-  // localization sweep. Role-based items (cut/copy/paste/reload/etc.) keep
-  // their OS-localized labels — those auto-translate when the user's system
-  // language matches; we only override the explicit `label` strings.
-  Menu.setApplicationMenu(
-    Menu.buildFromTemplate([
-      {
-        label: 'Maka',
-        submenu: [
-          { role: 'about', label: '关于 Maka' },
-          {
-            label: '设置…',
-            accelerator: 'CommandOrControl+,',
-            click: () => safeSendToRenderer('window:openSettings'),
-          },
-          { type: 'separator' },
-          { role: 'hide', label: '隐藏 Maka' },
-          { role: 'hideOthers' },
-          { role: 'unhide' },
-          { type: 'separator' },
-          { role: 'quit', label: '退出 Maka' },
-        ],
-      },
-      { label: '文件', submenu: [{ role: 'close' }] },
-      {
-        label: '编辑',
-        submenu: [
-          { role: 'undo' },
-          { role: 'redo' },
-          { type: 'separator' },
-          { role: 'cut' },
-          { role: 'copy' },
-          { role: 'paste' },
-          { role: 'selectAll' },
-        ],
-      },
-      {
-        label: '视图',
-        submenu: [
-          { role: 'reload' },
-          { role: 'toggleDevTools' },
-          { type: 'separator' },
-          { role: 'resetZoom' },
-          { role: 'zoomIn' },
-          { role: 'zoomOut' },
-          { type: 'separator' },
-          { role: 'togglefullscreen' },
-        ],
-      },
-      { label: '窗口', submenu: [{ role: 'minimize' }, { role: 'zoom' }] },
-    ]),
-  );
-}
-
 function localMemoryOpenFailureCopy(reason: string): string {
   switch (reason) {
     case 'incognito_blocked':
@@ -2088,11 +1718,7 @@ function registerIpc(): void {
   }
 
   ipcMain.handle('window:setTitlebarControlsVisible', (event, visible: unknown): void => {
-    const target = BrowserWindow.fromWebContents(event.sender);
-    if (!target || target !== mainWindow || process.platform !== 'darwin') return;
-    const shouldShow = visible === true;
-    target.setWindowButtonVisibility(shouldShow);
-    target.setWindowButtonPosition(shouldShow ? MAIN_WINDOW_TRAFFIC_LIGHT_POSITION : HIDDEN_TRAFFIC_LIGHT_POSITION);
+    mainWindowController.setTitlebarControlsVisible(event.sender, visible);
   });
   ipcMain.handle('app:info', async () => {
     const projectPath = await currentProjectRoot();
@@ -2124,15 +1750,10 @@ function registerIpc(): void {
       | { ok: true; projectPath: string; projectGit: Awaited<ReturnType<typeof resolveProjectGitInfo>> }
       | { ok: false; reason: 'cancelled' | 'missing-selection' }
     > => {
-      const result = mainWindow
-        ? await dialog.showOpenDialog(mainWindow, {
-            title: '选择工作目录',
-            properties: ['openDirectory'],
-          })
-        : await dialog.showOpenDialog({
-            title: '选择工作目录',
-            properties: ['openDirectory'],
-          });
+      const result = await mainWindowController.showOpenDialog({
+        title: '选择工作目录',
+        properties: ['openDirectory'],
+      });
       const selectedPath = result.filePaths[0];
       if (result.canceled) return { ok: false, reason: 'cancelled' };
       if (!selectedPath) return { ok: false, reason: 'missing-selection' };
@@ -2294,17 +1915,11 @@ function registerIpc(): void {
         { name: 'Office', extensions: ['docx', 'xlsx', 'pptx'] },
         { name: 'All Files', extensions: ['*'] },
       ];
-      const result = mainWindow
-        ? await dialog.showOpenDialog(mainWindow, {
-            title: '导入文件内容',
-            properties: ['openFile', 'multiSelections'],
-            filters: textFileFilters,
-          })
-        : await dialog.showOpenDialog({
-            title: '导入文件内容',
-            properties: ['openFile', 'multiSelections'],
-            filters: textFileFilters,
-          });
+      const result = await mainWindowController.showOpenDialog({
+        title: '导入文件内容',
+        properties: ['openFile', 'multiSelections'],
+        filters: textFileFilters,
+      });
       if (result.canceled || !result.filePaths[0]) {
         return { ok: false, reason: 'cancelled', message: '已取消导入。' };
       }
@@ -2346,15 +1961,10 @@ function registerIpc(): void {
       | { ok: false; reason: 'cancelled'; message: string }
       | { ok: false; reason: FolderOutlineImportFailureReason; message: string }
     > => {
-      const result = mainWindow
-        ? await dialog.showOpenDialog(mainWindow, {
-            title: '导入文件夹目录',
-            properties: ['openDirectory', 'multiSelections'],
-          })
-        : await dialog.showOpenDialog({
-            title: '导入文件夹目录',
-            properties: ['openDirectory', 'multiSelections'],
-          });
+      const result = await mainWindowController.showOpenDialog({
+        title: '导入文件夹目录',
+        properties: ['openDirectory', 'multiSelections'],
+      });
       if (result.canceled || !result.filePaths[0]) {
         return { ok: false, reason: 'cancelled', message: '已取消导入。' };
       }
@@ -2423,9 +2033,7 @@ function registerIpc(): void {
       title: `另存为 ${record.name}`,
       defaultPath: record.name,
     };
-    const result = mainWindow
-      ? await dialog.showSaveDialog(mainWindow, saveDialogOptions)
-      : await dialog.showSaveDialog(saveDialogOptions);
+    const result = await mainWindowController.showSaveDialog(saveDialogOptions);
     if (result.canceled || !result.filePath) return { ok: false, reason: 'canceled' };
     try {
       await copyFile(resolved.path, result.filePath);
@@ -2462,10 +2070,11 @@ function registerIpc(): void {
       const scenario = sanitizeSegment(input?.scenario);
       const variant = sanitizeSegment(input?.variant);
       if (!scenario || !variant) return { ok: false, reason: 'invalid_input' };
-      if (!mainWindow) return { ok: false, reason: 'capture_failed' };
       let image: Electron.NativeImage;
       try {
-        image = await mainWindow.webContents.capturePage();
+        const capture = await mainWindowController.capturePage();
+        if (!capture) return { ok: false, reason: 'capture_failed' };
+        image = capture;
       } catch {
         return { ok: false, reason: 'capture_failed' };
       }
@@ -3166,7 +2775,7 @@ function registerIpc(): void {
   // renderer channels.
   // The getter reads the live shownBrowserSessionId so the host's visible-lease
   // gate (canDrive) reflects the conversation the window currently shows.
-  provideBrowserViewHost(createBrowserViewHost(getBrowserViews(), () => shownBrowserSessionId));
+  provideBrowserViewHost(createBrowserViewHost(mainWindowController.getBrowserViews(), () => shownBrowserSessionId));
 
   // Never trust the renderer's target: it must be the session the calling
   // window currently shows (reported via browser:active-session). The agent
@@ -3178,7 +2787,7 @@ function registerIpc(): void {
     // stale one can never float over the newly-shown conversation, regardless of
     // renderer effect ordering or a reload. The shown view is re-positioned by
     // its panel's rect mirror.
-    getBrowserViews().hideAllExcept(shownBrowserSessionId);
+    mainWindowController.getBrowserViews().hideAllExcept(shownBrowserSessionId);
     // The visible lease is continuous: revoke any browser action still running
     // for a conversation that just went off screen, so it can't keep reading or
     // driving a hidden, logged-in page. canDrive only gates the START.
@@ -3191,24 +2800,24 @@ function registerIpc(): void {
   // native view tracks it; a null rect (modal open / panel unmounted) hides it.
   ipcMain.on('browser:setViewport', (_event, input: { sessionId?: unknown; rect?: BrowserViewRect | null }) => {
     if (!browserTargetOk(input?.sessionId)) return;
-    getBrowserViews().setViewport(input.sessionId, input.rect ?? null);
+    mainWindowController.getBrowserViews().setViewport(input.sessionId, input.rect ?? null);
   });
   // Create on first navigate so conversations that never open the browser pay nothing.
   ipcMain.handle('browser:navigate', async (_event, target: unknown, url: unknown) => {
     if (!browserTargetOk(target)) return;
-    await getBrowserViews().getOrCreate(target).navigate(String(url ?? ''));
+    await mainWindowController.getBrowserViews().getOrCreate(target).navigate(String(url ?? ''));
   });
   ipcMain.handle('browser:back', (_event, target: unknown) => {
-    if (browserTargetOk(target)) getBrowserViews().get(target)?.goBack();
+    if (browserTargetOk(target)) mainWindowController.getBrowserViews().get(target)?.goBack();
   });
   ipcMain.handle('browser:forward', (_event, target: unknown) => {
-    if (browserTargetOk(target)) getBrowserViews().get(target)?.goForward();
+    if (browserTargetOk(target)) mainWindowController.getBrowserViews().get(target)?.goForward();
   });
   ipcMain.handle('browser:reload', (_event, target: unknown) => {
-    if (browserTargetOk(target)) getBrowserViews().get(target)?.reload();
+    if (browserTargetOk(target)) mainWindowController.getBrowserViews().get(target)?.reload();
   });
   ipcMain.handle('browser:stop', (_event, target: unknown) => {
-    if (browserTargetOk(target)) getBrowserViews().get(target)?.stop();
+    if (browserTargetOk(target)) mainWindowController.getBrowserViews().get(target)?.stop();
   });
   // Read-only state query, intentionally NOT gated by browserTargetOk: the panel
   // issues it from its mount effect, which runs BEFORE the parent's
@@ -3218,7 +2827,9 @@ function registerIpc(): void {
   // is not a trust boundary — only mutation (navigate/back/...) and view
   // positioning (setViewport) are, and those stay guarded.
   ipcMain.handle('browser:get-state', (_event, target: unknown) =>
-    typeof target === 'string' && target.length > 0 ? (getBrowserViews().get(target)?.state() ?? null) : null,
+    typeof target === 'string' && target.length > 0
+      ? (mainWindowController.getBrowserViews().get(target)?.state() ?? null)
+      : null,
   );
   // The tab's × promises "Close": destroy the conversation's page outright via
   // the same dispose chain as session delete.
@@ -4604,7 +4215,7 @@ app.whenReady().then(async () => {
   await recoverInterruptedSessionsOnStartup();
   await botRegistry.applySettings(settings.botChat);
   await openGateway.sync(settings.openGateway);
-  await createWindow();
+  await mainWindowController.createWindow();
   await refreshPlanReminderTimers();
   startDailyReviewScheduler();
 });
@@ -4618,9 +4229,9 @@ app.on('before-quit', () => {
   if (dailyReviewSchedulerTimer) clearInterval(dailyReviewSchedulerTimer);
   void botRegistry.stopAll();
   void openGateway.stop();
-  void browserViews?.disposeAll();
+  void mainWindowController.disposeBrowserViews();
 });
 
 app.on('activate', () => {
-  if (BrowserWindow.getAllWindows().length === 0) void createWindow();
+  if (!mainWindowController.hasOpenWindows()) void mainWindowController.createWindow();
 });
